+++ CHANGES.rst
@@ -131,6 +131,7 @@ AudioLazy changes history
     as Sympy symbols (including symbolic matrices).
   - New highpass filter strategies ``highpass.z`` (taken as the new default)
     and ``highpass.pole_exp``
+  - New lowpass filter strategy ``lowpass.z``
 
 + lazy_io: